refactor: Shop 타입 구조 변경
