+++ src/hooks/api/shop/useGetShopInfoQuery.ts
@@ -1,8 +1,10 @@
-import { ShopResponse } from "@/types/api/shop";
 import { instance } from "@/utils/instance";
 import { useQuery } from "@tanstack/react-query";
+import { ShopResponse } from "./shop";
 
-const getShopInfo = async (shopId: string): Promise<ShopResponse> => {
+export type GetShopInfoResponse = ShopResponse;
+
+const getShopInfo = async (shopId: string): Promise<GetShopInfoResponse> => {
   const response = await instance.get(`/shops/${shopId}`);
   return response.data;
 };
