+++ src/hooks/api/shop/usePutShopInfoQuery.ts
@@ -1,8 +1,15 @@
 import { useMutation } from "@tanstack/react-query";
 import { instance } from "@/utils/instance";
-import { ShopRequest, ShopResponse } from "@/types/api/shop";
+import { ShopRequest, ShopResponse } from "./shop";
 
-const putShopInfo = async ({ shopId, data }: { shopId: string; data: ShopRequest }): Promise<ShopResponse> => {
+export interface PutShopInfoRequest {
+  shopId: string;
+  data: ShopRequest;
+}
+
+export type PutShopInfoResponse = ShopResponse;
+
+const putShopInfo = async ({ shopId, data }: PutShopInfoRequest): Promise<PutShopInfoResponse> => {
   const response = await instance.put(`/shops/${shopId}`, data);
   return response.data;
 };
